Move query condition API sources to `api` directory.
