+++ tiledb/sm/c_api/tiledb_experimental.h
@@ -49,6 +49,7 @@
 #include "tiledb/api/c_api/fragment_info/fragment_info_api_experimental.h"
 #include "tiledb/api/c_api/profile/profile_api_experimental.h"
 #include "tiledb/api/c_api/query_aggregate/query_aggregate_api_external_experimental.h"
+#include "tiledb/api/c_api/query_condition/query_condition_api_external_experimental.h"
 #include "tiledb/api/c_api/query_field/query_field_api_external_experimental.h"
 #include "tiledb/api/c_api/query_plan/query_plan_api_external_experimental.h"
 #include "tiledb/api/c_api/subarray/subarray_api_experimental.h"
@@ -147,77 +148,6 @@ TILEDB_EXPORT int32_t tiledb_query_get_relevant_fragment_num(
     tiledb_ctx_t* ctx,
     const tiledb_query_t* query,
     uint64_t* relevant_fragment_num) TILEDB_NOEXCEPT;
-
-/* ********************************* */
-/*          QUERY CONDITION          */
-/* ********************************* */
-
-/**
- * Initializes a TileDB query condition set membership object.
- *
- * **Example:**
- *
- * @code{.c}
- * tiledb_query_condition_t* cond
- * tiledb_query_condition_alloc_set_membership(
- *   ctx,
- *   "some_name",
- *   data,
- *   data_size,
- *   offsets,
- *   offsets_size,
- *   TILEDB_QUERY_CONDITION_OP_IN,
- *   &cond);
- * @endcode
- *
- * @param ctx The TileDB context.
- * @param field_name The field name.
- * @param data A pointer to the set member data.
- * @param data_size The length of the data buffer.
- * @param offsets A pointer to the array of offsets of members.
- * @param offsets_size The length of the offsets array in bytes.
- * @param op The set membership operator to use.
- * @param cond The allocated query condition object.
- * @return `TILEDB_OK` for success and `TILEDB_ERR` for error.
- */
-TILEDB_EXPORT capi_return_t tiledb_query_condition_alloc_set_membership(
-    tiledb_ctx_t* ctx,
-    const char* field_name,
-    const void* data,
-    uint64_t data_size,
-    const void* offsets,
-    uint64_t offests_size,
-    tiledb_query_condition_op_t op,
-    tiledb_query_condition_t** cond) TILEDB_NOEXCEPT;
-
-/**
- * Disable the use of enumerations on the given QueryCondition
- *
- * **Example:**
- *
- * @code{.c}
- * tiledb_query_condition_t* query_condition;
- * tiledb_query_condition_alloc(ctx, &query_condition);
- * uint32_t value = 5;
- * tiledb_query_condition_init(
- *   ctx,
- *   query_condition,
- *   "longitude",
- *   &value,
- *   sizeof(value),
- *   TILEDB_LT);
- * tiledb_query_condition_set_use_enumeration(ctx, query_condition, 0);
- * @endcode
- *
- * @param[in] ctx The TileDB context.
- * @param[in] cond The query condition
- * @param[in] use_enumeration Non-zero to use the associated enumeration
- * @return `TILEDB_OK` for success and `TILEDB_ERR` for error.
- */
-TILEDB_EXPORT int32_t tiledb_query_condition_set_use_enumeration(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_condition_t* cond,
-    int use_enumeration) TILEDB_NOEXCEPT;
 
 /* ********************************* */
 /*        QUERY STATUS DETAILS       */
